@@ -1,22 +1,22 @@
 name: stash_disk
-version: 1.0.2
+version: 2.0.0-nullsafety.1
 description: >-
   Disk storage extension for the stash caching API. Provides support to store caches
   in the local file system in binary format using the msgpack json format
 homepage: https://github.com/ivoleitao/stash_disk
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  path: ^1.7.0
-  file: ^5.2.1
-  stream_transform: ^1.2.0
-  stash: ^1.0.2
-  #stash:
-  #  path: ../stash
+  path: ^1.8.0
+  file: ^6.1.0
+  stream_transform: ^2.0.0
+  # stash: ^1.0.2
+  stash:
+    path: ../stash
 
 dev_dependencies:
-  pedantic: ^1.9.2
-  test: ^1.15.4
-  test_coverage: ^0.4.3
+  pedantic: ^1.10.0
+  test: ^1.16.4
+  #test_coverage: ^0.4.3
